+++ pygal/graph/public.py
@@ -80,6 +80,7 @@ class PublicApi(BaseGraph):
             from lxml.html import open_in_browser
         except ImportError:
             raise ImportError('You must install lxml to use render in browser')
+        kwargs.setdefault('force_uri_protocol', 'https')
         open_in_browser(self.render_tree(**kwargs), encoding='utf-8')
 
     def render_response(self, **kwargs):